// sanity/types.ts
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export type PROJECT_BY_ID_QUERYResult = {
  _id: string;
  title: string | null;
  slug: Slug | null;
  _createdAt: string;
  author: {
    _id: string;
    name: string | null;
    username: string | null;
    image: string | null;
    bio: string | null;
  } | null;
  views: number | null;
  description: string | null;
  category: string | null;
  image: string | null;
  image2: string | null;
  image3: string | null;
  pitch: string | null;
  techStack: Array<string> | null;
  githubUrl: string | null;
  upvotes: number | null;
  comments: Array<{
    _id: string;
    text: string | null;
    _createdAt: string;
    author: {
      _id: string;
      name: string | null;
      image: string | null;
    } | null;
  }>;
} | null;